refactor: move helper functions to new file telemetryUtils.ts
