+++ packages/salesforcedx-utils-vscode/src/helpers/index.ts
@@ -42,3 +42,4 @@ export {
 } from './utils';
 export { isAlphaNumSpaceString, isAlphaNumString, isInteger, isIntegerInRange, isRecordIdFormat } from './validations';
 export { errorToString, isError, isErrorString, getErrorMessage, getErrorStack } from './errorUtils';
+export { getSharedTelemetryUserId, hashUserIdentifier, refreshAllExtensionReporters } from './telemetryUtils';
